Resolve a circular dependency in collections
