+++ frontend/src/metabase/collections/components/ModelUploadModal.tsx
@@ -12,25 +12,12 @@ import {
   Stack,
   Text,
 } from "metabase/ui";
-import type { CardId, CollectionId, TableId } from "metabase-types/api";
+import type { CollectionId, TableId } from "metabase-types/api";
 import { UploadMode } from "metabase-types/store/upload";
 
 import type { OnFileUpload } from "../types";
 
 import { findLastEditedCollectionItem } from "./utils";
-
-export type CollectionOrTableIdProps =
-  | {
-      uploadMode: UploadMode.create;
-      collectionId: CollectionId;
-      tableId?: never;
-    }
-  | {
-      uploadMode: UploadMode.append | UploadMode.replace;
-      collectionId?: never;
-      tableId: TableId;
-      modelId?: CardId;
-    };
 
 export function ModelUploadModal({
   opened,
